Move implements from exapmles to impl
