+++ examples/example_allegro5/main.cpp
@@ -5,7 +5,7 @@
 #include <allegro5/allegro.h>
 #include <allegro5/allegro_primitives.h>
 #include "imgui.h"
-#include "imgui_impl_allegro5.h"
+#include "impl/imgui_impl_allegro5.h"
 
 int main(int, char**)
 {
