+++ examples/example_glfw_opengl2/main.cpp
@@ -7,8 +7,8 @@
 // See imgui_impl_glfw.cpp for details.
 
 #include "imgui.h"
-#include "imgui_impl_glfw.h"
-#include "imgui_impl_opengl2.h"
+#include "impl/imgui_impl_glfw.h"
+#include "impl/imgui_impl_opengl2.h"
 #include <stdio.h>
 #include <GLFW/glfw3.h>
