+++ examples/example_marmalade/main.cpp
@@ -5,7 +5,7 @@
 // This file is part of Dear ImGui
 
 #include "imgui.h"
-#include "imgui_impl_marmalade.h"
+#include "impl/imgui_impl_marmalade.h"
 #include <stdio.h>
 
 #include <s3eKeyboard.h>
