+++ examples/example_sdl_opengl2/main.cpp
@@ -7,8 +7,8 @@
 // See imgui_impl_sdl.cpp for details.
 
 #include "imgui.h"
-#include "imgui_impl_sdl.h"
-#include "imgui_impl_opengl2.h"
+#include "impl/imgui_impl_sdl.h"
+#include "impl/imgui_impl_opengl2.h"
 #include <stdio.h>
 #include <SDL.h>
 #include <SDL_opengl.h>
