+++ examples/example_sdl_opengl3/main.cpp
@@ -4,8 +4,8 @@
 // (GL3W is a helper library to access OpenGL functions since there is no standard header to access modern OpenGL functions easily. Alternatives are GLEW, Glad, etc.)
 
 #include "imgui.h"
-#include "imgui_impl_sdl.h"
-#include "imgui_impl_opengl3.h"
+#include "impl/imgui_impl_sdl.h"
+#include "impl/imgui_impl_opengl3.h"
 #include <stdio.h>
 #include <SDL.h>
