+++ impl/imgui_impl_dx10.cpp
@@ -18,7 +18,7 @@
 //  2018-02-06: Misc: Removed call to ImGui::Shutdown() which is not available from 1.60 WIP, user needs to call CreateContext/DestroyContext themselves.
 //  2016-05-07: DirectX10: Disabling depth-write.
 
-#include "imgui.h"
+#include "../imgui.h"
 #include "imgui_impl_dx10.h"
 
 // DirectX
