+++ impl/imgui_impl_marmalade.cpp
@@ -14,7 +14,7 @@
 //  2018-02-06: Misc: Removed call to ImGui::Shutdown() which is not available from 1.60 WIP, user needs to call CreateContext/DestroyContext themselves.
 //  2018-02-06: Inputs: Added mapping for ImGuiKey_Space.
 
-#include "imgui.h"
+#include "../imgui.h"
 #include "imgui_impl_marmalade.h"
 
 #include <s3eClipboard.h>
