+++ impl/imgui_impl_sdl.cpp
@@ -33,7 +33,7 @@
 //  2017-08-25: Inputs: MousePos set to -FLT_MAX,-FLT_MAX when mouse is unavailable/missing (instead of -1,-1).
 //  2016-10-15: Misc: Added a void* user_data parameter to Clipboard function handlers.
 
-#include "imgui.h"
+#include "../imgui.h"
 #include "imgui_impl_sdl.h"
 
 // SDL
