+++ impl/imgui_impl_vulkan.cpp
@@ -26,7 +26,7 @@
 //  2016-10-18: Vulkan: Add location decorators & change to use structs as in/out in glsl, update embedded spv (produced with glslangValidator -x). Null the released resources.
 //  2016-08-27: Vulkan: Fix Vulkan example for use when a depth buffer is active.
 
-#include "imgui.h"
+#include "../imgui.h"
 #include "imgui_impl_vulkan.h"
 #include <stdio.h>
